warn user about actions on SIB

+++ workspace/it.unibo.games.rpsls.admin/src/it/unibo/games/rpsls/admin/gui/ViewMain.java
@@ -16,11 +16,14 @@ import javax.swing.BoxLayout;
 import javax.swing.JButton;
 import javax.swing.JLabel;
 import javax.swing.JList;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.ListSelectionModel;
+import javax.swing.event.ListSelectionEvent;
+import javax.swing.event.ListSelectionListener;
 
-public class ViewMain extends ViewDefault implements ActionListener {
+public class ViewMain extends ViewDefault implements ActionListener, ListSelectionListener {
 
 	private HashMap<String, IGame> games;
 	private JList<String> list;
@@ -72,11 +75,13 @@ public class ViewMain extends ViewDefault implements ActionListener {
 		main.setAlignmentY(TOP_ALIGNMENT);
 		list = new JList<String>();
 		list.setSelectionMode(ListSelectionModel.MULTIPLE_INTERVAL_SELECTION);
+		list.addListSelectionListener(this);
 //		list.addMouseListener(this);
 		JScrollPane pL = new JScrollPane(list);
 		JPanel p1 = new JPanel();
 		deleteGame = new JButton("Delete");
-
+		deleteGame.setEnabled(false);
+		
 		p1.add(deleteGame);
 		main.add(pL);
 		main.add(p1);	
@@ -93,8 +98,8 @@ public class ViewMain extends ViewDefault implements ActionListener {
 
 		pB1.add(adminButtons);
 		pB1.add(init);
-		pB1.add(reset);
 		pB1.add(clean);
+		pB1.add(reset);
 		adminButtons.setAlignmentY(BOTTOM_ALIGNMENT);
 			
 		pB.add(pB1);
@@ -108,19 +113,35 @@ public class ViewMain extends ViewDefault implements ActionListener {
 		clean.addActionListener(this);
 		
 	}
+	
+	private boolean showDialog(String label) {
+		int dialogButton = JOptionPane.YES_NO_OPTION;
+		int dialogResult = JOptionPane.showConfirmDialog(this, label, "Confirm action", dialogButton, JOptionPane.PLAIN_MESSAGE);
+		return dialogResult == 0;
+	}
 
 	@Override
 	public void actionPerformed(ActionEvent arg0) {
 		JButton source = (JButton) arg0.getSource();
 		
-		if (source == deleteGame)
-			deleteGame();
-		else if (source == init)
+		if (source == deleteGame) {
+			int games_num = list.getSelectedValuesList().size();
+			if (games_num > 0 && showDialog("You are going to remove " + games_num + " games from SIB. Are you sure about that?"))
+				deleteGame();
+		}
+
+		else if (source == init && showDialog("You are going to inject ontology into SIB. Are you sure about that?"))
 			mainWindow.connector.init();
-		else if (source == reset)
+		else if (source == reset && showDialog("You are going to remove everything from SIB. Are you sure about that?"))
 			mainWindow.connector.reset();
-		else if (source == clean)
+		else if (source == clean && showDialog("You are going to remove every RPSLS data from SIB. Are you sure about that?"))
 			mainWindow.connector.clean();
+	}
+
+	@Override
+	public void valueChanged(ListSelectionEvent arg0) {
+		int games_num = list.getSelectedValuesList().size();
+		deleteGame.setEnabled(games_num > 0);
 	}
